// src/types/admin-all-data-get-type.ts
export interface getdataTypes {
  token_id: string;
  userObj: any;
  created_on: string;
  _id: number;
  is_hide: boolean;
  title: string;
  description: string;
  collectible_owner: string;
  collection_address: string;
  collection_status: string;
  last_auction_type: string;
  last_buy_currency: any;
  on_sale: boolean;
  network_id: string;
  collectible_id: {
    wallet_addres: string;
  };
}